Add an integration test
Here, we start a container running a PyPI-compatible service
locally. We also build a small Python package to use for uploading.
If we're able to use the plugin to upload the package, the command
hook should succeed, and the `pip search` from the post-command hook
should also succeed.

Because of the networking requirements, the `twine` shim we have for
running in a Docker container is conditionally "tweaked" only when
running in the integration test job of this repository's verify
pipeline to add the necessary `host.docker.internal` hostname.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>